refactor: reorganize rest api folder
