+++ src/api/index.ts
@@ -1,6 +1,6 @@
-import * as exportList from '~api/rest/exportList';
-import * as lists from '~api/rest/lists';
-import * as persons from '~api/rest/persons';
+import * as exportList from '~api/rest/exportList/exportList';
+import * as lists from '~api/rest/lists/lists';
+import * as persons from '~api/rest/persons/persons';
 
 export const api = {
 	exportList,
