+++ src/api/rest/persons/persons.ts
@@ -3,4 +3,7 @@ import type { PersonRequest } from '~api/types/backend/requests/PersonRequest';
 import type { ListsResponse } from '~api/types/backend/responses/ListsResponse';
 
 export const addNewPerson = (listId: string, body: PersonRequest) =>
-	commonPatch<ListsResponse, PersonRequest>(`/order/${listId}`, body);
+	commonPatch<ListsResponse, PersonRequest>(`/order/${listId}`, body).then((resp) => {
+		console.log(resp);
+		return resp;
+	});
